+++ src/cmd_spanr/coverage.rs
@@ -10,22 +10,28 @@ type Iv = Interval<u32, u32>; // the first type should be Unsigned
 // Create clap subcommand arguments
 pub fn make_subcommand<'a>() -> Command<'a> {
     Command::new("coverage")
-        .about("Output detailed depths of coverages on chromosomes")
+        .about("Output minimum or detailed depth of coverage on chromosomes")
         .arg(
             Arg::new("infiles")
-                .help("Sets the input file to use")
+                .help("Set the input file to use")
                 .required(true)
                 .min_values(1)
                 .index(1),
         )
         .arg(
-            Arg::new("coverage")
-                .help("minimal coverage")
-                .long("coverage")
-                .short('c')
+            Arg::new("minimum")
+                .help("Set the minimum depth of coverage")
+                .long("minimum")
+                .short('m')
                 .takes_value(true)
                 .default_value("1")
                 .forbid_empty_values(true),
+        )
+        .arg(
+            Arg::new("detailed")
+                .help("Output detailed depth")
+                .long("detailed")
+                .short('d'),
         )
         .arg(
             Arg::new("outfile")
@@ -43,10 +49,11 @@ pub fn execute(args: &ArgMatches) -> std::result::Result<(), std::io::Error> {
     //----------------------------
     // Loading
     //----------------------------
-    let coverage: i32 = args.value_of_t("coverage").unwrap_or_else(|e| {
-        eprintln!("Need a integer for --coverage\n{}", e);
+    let minimum: i32 = args.value_of_t("minimum").unwrap_or_else(|e| {
+        eprintln!("Need a integer for --minimum\n{}", e);
         std::process::exit(1)
     });
+    let is_detailed = args.is_present("detailed");
 
     // seq_name => Vector of Intervals
     let mut res: BTreeMap<String, Vec<Iv>> = BTreeMap::new();
@@ -77,25 +84,73 @@ pub fn execute(args: &ArgMatches) -> std::result::Result<(), std::io::Error> {
     //----------------------------
     // Output
     //----------------------------
-    let mut set: BTreeMap<String, IntSpan> = BTreeMap::new();
-    for chr in res.keys() {
-        let lapper = Lapper::new(res.get(chr).unwrap().to_owned());
-        let ivs = lapper.depth().collect::<Vec<Interval<u32, u32>>>();
+    if is_detailed {
+        // Multi
+        let mut set_of: BTreeMap<String, BTreeMap<String, IntSpan>> = BTreeMap::new();
 
-        let mut intspan = IntSpan::new();
-        for iv in ivs {
-            let depth = iv.val as i32;
-            if depth < coverage {
-                continue;
+        for chr in res.keys() {
+            let lapper = Lapper::new(res.get(chr).unwrap().to_owned());
+            let ivs = lapper.depth().collect::<Vec<Interval<u32, u32>>>();
+
+            // depth => IntSpan
+            let mut intspan_of: BTreeMap<String, IntSpan> = BTreeMap::new();
+
+            for iv in ivs {
+                let depth = iv.val as i32;
+                if depth < minimum {
+                    continue;
+                }
+
+                let depth = format!("{}", depth);
+
+                if !set_of.contains_key(&depth) {
+                    set_of.insert(depth.clone(), BTreeMap::new());
+                }
+
+                if !intspan_of.contains_key(&depth) {
+                    intspan_of.insert(depth.clone(), IntSpan::new());
+                }
+
+                intspan_of
+                    .entry(depth)
+                    .and_modify(|e| e.add_pair(iv.start as i32, iv.stop as i32 - 1));
             }
 
-            intspan.add_pair(iv.start as i32, iv.stop as i32 - 1);
+            for depth in intspan_of.keys() {
+                set_of
+                    .get_mut(depth)
+                    .unwrap()
+                    .insert(chr.clone(), intspan_of.get(depth).unwrap().clone());
+            }
         }
 
-        set.insert(chr.to_string(), intspan);
+        let out_yaml = set2yaml_m(&set_of);
+        write_yaml(args.value_of("outfile").unwrap(), &out_yaml)?;
+    } else {
+        // Single
+        // chr => IntSpan
+        let mut set: BTreeMap<String, IntSpan> = BTreeMap::new();
+
+        for chr in res.keys() {
+            let lapper = Lapper::new(res.get(chr).unwrap().to_owned());
+            let ivs = lapper.depth().collect::<Vec<Interval<u32, u32>>>();
+
+            let mut intspan = IntSpan::new();
+            for iv in ivs {
+                let depth = iv.val as i32;
+                if depth < minimum {
+                    continue;
+                }
+
+                intspan.add_pair(iv.start as i32, iv.stop as i32 - 1);
+            }
+
+            set.insert(chr.to_string(), intspan);
+        }
+
+        let out_yaml = set2yaml(&set);
+        write_yaml(args.value_of("outfile").unwrap(), &out_yaml)?;
     }
-    let out_yaml = set2yaml(&set);
-    write_yaml(args.value_of("outfile").unwrap(), &out_yaml)?;
 
     Ok(())
 }